refactor(avro): keep AppendDefaultToBuilder local to Avro
Row-oriented Avro decode needs per-builder appends, while Parquet uses
batch MakeDefaultArray. Share ToArrowScalar only; don't put the Avro
shape into the shared arrow literal util.
